[dm] [cli] fixed test outputs
