+++ cli/Makefile
@@ -14,4 +14,4 @@ build:
 
 .PHONY: test
 test:
-	go test ./cmd/...
+	go test ./...
