+++ cli/deployment/deployment.go
@@ -1,0 +1,6 @@
+package deployment
+
+type Deployment struct {
+	name    string
+	outputs map[string]string
+}
