--- cli/dm_utils.go
@@ -1,9 +1,0 @@
-package main
-
-import "hash/fnv"
-
-func hash64(s string) int64 {
-	hash := fnv.New32()
-	hash.Write([]byte(s))
-	return int64(hash.Sum32())
-}
